Qualified or Not Qualified
Write a program that reads the marks M in Maths and marks P in Physics and checks if both M and P are greater than 35 or sum of M and P is greater than or equal to 100.
Print Qualified if both M and P are greater than 35 or sum of M and P is greater than or equal to 100. Otherwise, print Not Qulified.